+++ src/lib/components/Select/Select.tsx
@@ -1,0 +1,168 @@
+import React, { KeyboardEventHandler, useEffect, useRef, useState } from 'react';
+import { Backdrop } from '../Backdrop';
+import { Checkbox } from '../Checkbox';
+import { Chip } from '../Chip';
+import { COLOR } from '../component.enums';
+import { List, ListItem, ListItemText } from '../List';
+import styles from './Select.module.scss';
+
+// TODO
+// navigate by keys
+// on key down enter
+
+export interface ISelectOption {
+	value: string;
+	label?: string;
+}
+
+export interface ISelectProps {
+	id?: string;
+	name?: string;
+	className?: string;
+	options: Array<ISelectOption>;
+	value?: string | Array<string>;
+	multiple?: boolean;
+	disabled?: boolean;
+	onChange?: (val: string | Array<string>) => void;
+}
+
+export const Select = (props: ISelectProps) => {
+
+	const { id, className, options, value, multiple, onChange } = props;
+
+	const [model, setModel] = useState<string | Array<string>>('');
+	const [isShow, setIsShow] = useState<boolean>(false);
+	const [selectedOptions, setSelectedOptions] = useState<Array<ISelectOption>>([]);
+	const selectConainter = useRef<HTMLDivElement>(null);
+
+	const getCssClass = () => {
+		const result = [];
+		result.push(className);
+		result.push(styles.select);
+		return result.filter(r => r).join(' ');
+	}
+
+	useEffect(() => {
+		const newValue = !!value ? value : '';
+		writeValue(newValue);
+	}, [value]);
+
+	const writeValue = (val: string | Array<string>) => setModel(val);
+
+	useEffect(() => {
+		if (!multiple) {
+			const newOption = options.find(o => o.value === model);
+			if (newOption) {
+				setSelectedOptions([newOption]);
+			}
+		} else {
+			const filteredOptions = options.filter(o => model.indexOf(o.value) >= 0);
+			setSelectedOptions([...filteredOptions]);
+		}
+	}, [model, multiple]);
+
+	const handleOnClick = (option: ISelectOption) => {
+		let newValue: string | Array<string> = multiple ? [] : '';
+
+		if (!multiple) {
+			if (model !== option.value) {
+				newValue = option.value;
+				onChange && onChange(newValue);
+			}
+			hide();
+		} else {
+			const selectedOption = selectedOptions.find(o => o.value === option.value);
+			if (selectedOption) {
+				newValue = selectedOptions.filter(o => o.value !== option.value).map(o => o.value);
+			} else {
+				newValue = (newValue as Array<string>).concat(selectedOptions.map(o => o.value));
+				(newValue as Array<string>).push(option.value);
+			}
+			onChange && onChange(newValue);
+		}
+
+		writeValue(newValue);
+	}
+
+	const show = () => setIsShow(true);
+	const hide = () => setIsShow(false);
+	const isActive = (option: ISelectOption) => selectedOptions.indexOf(option) >= 0;
+
+	const renderSingleViewModel = () => {
+		let result = null;
+		if (selectedOptions.length > 0) {
+			result = <span>{selectedOptions[0].label}</span>;
+		}
+		return result;
+	}
+
+	const renderMultipleViewModel = () => {
+		let result = null;
+		if (selectedOptions.length > 0) {
+			result = selectedOptions
+				.map(o =>
+					<Chip color={COLOR.primary} key={o.value} className="mr-2" onDelete={(e) => handleOnDelete((e as any), o)}>
+						{o.label}
+					</Chip>
+				);
+		}
+		return result;
+	}
+
+	const handleOnDelete = (event: Event, option: ISelectOption) => {
+		event.stopPropagation();
+		handleOnClick(option);
+	}
+
+	// TODO
+	const handleOnKeyDown = (e: KeyboardEventHandler<HTMLDivElement>) => {
+		if ((e as any).key === 'Enter') {
+		}
+	}
+
+	return (
+		<>
+			<div ref={selectConainter} className={styles.selectContainer}>
+
+				<div id={id} className={getCssClass()} onClick={() => show()} tabIndex={0} onKeyDown={e => handleOnKeyDown(e as any)}>
+					{!multiple && renderSingleViewModel()}
+
+					{multiple && renderMultipleViewModel()}
+
+					{/* {multiple &&
+						selectedOptions
+							.map(o =>
+								<Chip color={COLOR.primary} key={o.value} className="mr-2" onDelete={(e) => handleOnDelete((e as any), o)}>
+									{o.label}
+								</Chip>
+							)
+					} */}
+				</div>
+
+				{isShow &&
+					<>
+						<div className={styles.selectMenu}>
+							<List>
+								{options && options.map((option) =>
+									<ListItem key={option.value} onClick={() => handleOnClick(option)} active={isActive(option)}>
+
+										{multiple &&
+											<Checkbox
+												checked={isActive(option)}
+												onChange={() => handleOnClick(option)}
+											/>
+										}
+
+										<ListItemText primary={option.label ? option.label : option.value} />
+									</ListItem>
+								)}
+							</List>
+						</div>
+						<Backdrop target={selectConainter.current as HTMLElement} isTransparent onClick={() => hide()} />
+					</>
+				}
+
+			</div>
+		</>
+	);
+}